Install dev-related packages from atom_devbox.yml
This commit adds a new list of packages (`devbox_packages`) meant to be
installed only when targeting the development box. This is to avoid
installing dev-related packages in production environments.

The tasks in `devbox.yml` are only executed when targeting Ubuntu hosts
since that's the only distro that we are testing when provisioning our
Vagrant box.

Sphinx-related requirements are pulled from atom-docs/requirements.txt.

Connects to https://github.com/artefactual-labs/ansible-atom/issues/78.