Use separate php-rh-8.yml task file for Rocky/RH
It makes easier to maintain the php tasks for RedHat. We already had a
task file only for php7.4 and RedHat. It is the same php-rh-7.4.yml file
but adding a task to remove old php7 (required for 2.9 upgrades from
2.7/2.8).

To avoid installing recommended packages, the following options have
been added to yum task:

   * allowerasing: true
   * install_weak_deps: false