Add support for AtoM 2.9 and PHP 8.x (#128)
* Add AtoM2.9 support

* Add Ubuntu 24.04 support

* Add AtoM 2.9 updates

- Add 2.9-search.yml (copy of 2.8)
- Bump AtoM npm version

* Update devbox packages

* Update devbox docs tasks

* Update README

* Use separate php-rh-8.yml task file for Rocky/RH

It makes easier to maintain the php tasks for RedHat. We already had a
task file only for php7.4 and RedHat. It is the same php-rh-7.4.yml file
but adding a task to remove old php7 (required for 2.9 upgrades from
2.7/2.8).

To avoid installing recommended packages, the following options have
been added to yum task:

   * allowerasing: true
   * install_weak_deps: false

* Use nodesource repo depending on OS

* Added  conditional for default npm version based on OS
* Use nodesource 22.x repo for Ubuntu >= 22.04
* Use nodesource 22.x repo for RedHat >= 8 and php >= 8.1
* Fix nodesource 14.x repo for Ubuntu 20.04

* Update README.md

Update README.md for AtoM 2.9 and new supported Operative Systems.

---------

Co-authored-by: Miguel Medinilla <mamedin@artefactual.com>
Co-authored-by: Douglas Cerna (Soy Douglas) <douglascerna@yahoo.com>